Move code to forms module
